Extract the CommandArgs implement class field
